+++ spec/dummy/config/routes.rb
@@ -1,8 +1,11 @@
 Rails.application.routes.draw do
   # Define your application routes per the DSL in https://guides.rubyonrails.org/routing.html
 
+
   # Defines the root path route ("/")
   root "home#show"
 
-  post '/', to: "home#submit"
+  # Allows another route, to test Turbo navigation.
+  get "/*p", to: "home#show"
+  post '/*p', to: "home#submit"
 end